+++ CHANGELOG.md
@@ -33,13 +33,13 @@
 
 #### Bug修复
 
-- (无)
+- 修改在常规配置下 ServiceExceptionCodes.UNDEFINE 的错误代码为 0 的bug。
 
 #### 功能移除
 
 - ~~删除（重命名）com.dwarfeng.subgrade.stack.dao.ConstraintLookupDao~~
 
-  该在创建之时已经做出注释警告用户不要使用，因此该类的删除不会造成太大的不兼容。
+  该在创建之时已经做出注释警告用户不要使用，因此该类的删除不会造成过度的不兼容。
 
 ---